+++ public_html/index.php
@@ -57,6 +57,7 @@ $optionalNewClassFiles = [
     APP_PATH . '/Services/Ads/GoogleAdsAPI.php',
     APP_PATH . '/Services/Ads/AdPiiHasher.php',
     APP_PATH . '/Jobs/SendAdConversionJob.php',
+    APP_PATH . '/Jobs/SendBookingConfirmationJob.php',
     APP_PATH . '/Services/OAuth/MetaOAuthClient.php',
     APP_PATH . '/Models/SubscriptionPlan.php',
     APP_PATH . '/Models/WalletTransaction.php',